+++ commercetools/commercetools-sdk-java-api/src/main/java-predicates-generated/com/commercetools/api/predicates/query/cart/CartSetDirectDiscountsIgnoreCartDiscountsActionQueryBuilderDsl.java
@@ -1,0 +1,28 @@
+
+package com.commercetools.api.predicates.query.cart;
+
+import com.commercetools.api.predicates.query.*;
+
+public class CartSetDirectDiscountsIgnoreCartDiscountsActionQueryBuilderDsl {
+    public CartSetDirectDiscountsIgnoreCartDiscountsActionQueryBuilderDsl() {
+    }
+
+    public static CartSetDirectDiscountsIgnoreCartDiscountsActionQueryBuilderDsl of() {
+        return new CartSetDirectDiscountsIgnoreCartDiscountsActionQueryBuilderDsl();
+    }
+
+    public StringComparisonPredicateBuilder<CartSetDirectDiscountsIgnoreCartDiscountsActionQueryBuilderDsl> action() {
+        return new StringComparisonPredicateBuilder<>(
+            BinaryQueryPredicate.of().left(new ConstantQueryPredicate("action")),
+            p -> new CombinationQueryPredicate<>(p,
+                CartSetDirectDiscountsIgnoreCartDiscountsActionQueryBuilderDsl::of));
+    }
+
+    public BooleanComparisonPredicateBuilder<CartSetDirectDiscountsIgnoreCartDiscountsActionQueryBuilderDsl> directDiscountsIgnoreCartDiscounts() {
+        return new BooleanComparisonPredicateBuilder<>(
+            BinaryQueryPredicate.of().left(new ConstantQueryPredicate("directDiscountsIgnoreCartDiscounts")),
+            p -> new CombinationQueryPredicate<>(p,
+                CartSetDirectDiscountsIgnoreCartDiscountsActionQueryBuilderDsl::of));
+    }
+
+}